+++ src/components/chat/content/index.scss
@@ -1,0 +1,81 @@
+@import "../../../assets/styles/colors.scss";
+@import "../../../assets/styles/variables.scss";
+
+.chat-content {
+  position: relative;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 14px;
+  width: 100%;
+  height: 100vh;
+  padding: 0 20px;
+  overflow-y: auto;
+}
+
+.chat-content__header {
+  position: sticky;
+  top: 0;
+  left: 0;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  height: 54px;
+  padding: 10px 0;
+  border-bottom: 1px solid $grey-lighten-2;
+  background-color: white;
+}
+
+.chat-content__info {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+}
+
+.chat-content__info__title {
+  font-weight: 600;
+  font-size: 13px;
+}
+
+.chat-content__footer {
+  position: sticky;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 11px;
+  height: 60px;
+  padding: 11px 0;
+  border-top: 1px solid $grey-lighten-2;
+  background-color: white;
+}
+
+.chat-content__messages {
+  display: flex;
+  flex-direction: column;
+  gap: 30px;
+}
+
+.chat-content__messages__date {
+  display: flex;
+  justify-content: center;
+  color: $grey;
+  font-size: 12px;
+}
+
+.chat-content__messages__item {
+  display: flex;
+
+  &_right {
+    justify-content: end;
+  }
+}
+
+.chat-content__notify {
+  display: flex;
+  justify-content: center;
+  color: $grey;
+}
