+++ src/controllers/AuthController.ts
@@ -1,19 +1,3 @@
-type UserLogin = {
-  login:string
-  password:string
-}
-
-type UserRegistration = {
-  email: string
-  login: string
-  first_name: string
-  second_name: string
-  phone: string
-  password: string
-  password_repeat: string
-}
-
-
 export class AuthController {
   static login(data: UserLogin) {
     console.log(data)
