+++ src/pages/main/index.scss
@@ -7,112 +7,3 @@
   height: 100vh;
   overflow: hidden;
 }
-
-.chat-sidebar {
-  display: flex;
-  flex-direction: column;
-  gap: 14px;
-  min-width: 310px;
-  width: 310px;
-  height: 100%;
-  background-color: #fbfbfb;
-  border-right: 1px solid $grey-lighten-2;
-}
-
-.chat-sidebar__header {
-  display: flex;
-  flex-direction: column;
-  height: 100px;
-  padding: 20px 10px 0;
-  gap: 20px;
-}
-
-.chat-sidebar__profile {
-  display: flex;
-  justify-content: end;
-}
-
-.chat-sidebar__chats {
-  display: flex;
-  flex-direction: column;
-  height: 100%;
-  overflow: hidden;
-  overflow-y: auto;
-
-  padding-bottom: 55px;
-}
-
-// hide scrollbar but still scrollable
-.chat-sidebar__chats {
-  -ms-overflow-style: none; /* Internet Explorer 10+ */
-  scrollbar-width: none; /* Firefox */
-}
-.chat-sidebar__chats::-webkit-scrollbar {
-  display: none; /* Safari and Chrome */
-}
-
-.chat-content {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  gap: 14px;
-  width: 100%;
-  height: 100vh;
-}
-
-.chat-content__header {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 60px;
-  height: 54px;
-  border-bottom: 1px solid $grey-lighten-2;
-  margin: 0 20px;
-}
-
-.chat-content__info {
-  display: flex;
-  align-items: center;
-  gap: 10px;
-}
-
-.chat-content__info__title {
-  font-weight: 600;
-  font-size: 13px;
-}
-
-.chat-content__footer {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 11px;
-  height: 60px;
-  margin: 0 20px;
-  padding: 11px 0;
-  border-top: 1px solid $grey-lighten-2;
-}
-
-.chat-content__messages {
-  display: flex;
-  flex-direction: column;
-  gap: 30px;
-  height: 100%;
-  padding: 0 20px;
-  overflow: hidden;
-  overflow-y: auto;
-}
-
-.chat-content__messages__date {
-  display: flex;
-  justify-content: center;
-  color: $grey;
-  font-size: 12px;
-}
-
-.chat-content__messages__item {
-  display: flex;
-
-  &_right {
-    justify-content: end;
-  }
-}
